Updates to all files
Mian has had several updaes, it now prints the HTML graph in consistent colours and has disabled the mouseover function (which was annoying). The format of the sentiment analysis has also been improved. Finaly main now displays the resilts compared to it's last run, by saving results out as text files, reading them back in and calculating the difference.

Vaid generator was failing when a username wasn't found - it now skips that error and write the not found user to a file. It also now properly read the time file back into memory on subsiquent runs.